requirements.txt
django
djangorestframework
pytest
pytest-django
pytest-cov
